Move TLS support into net/tls module
