+++ src/net.rs
@@ -10,9 +10,11 @@ use tokio::time::timeout;
 use tokio_io_timeout::TimeoutStream;
 
 use crate::context::Context;
+pub use crate::net::tls::wrap_tls;
 use crate::tools::time;
 
 pub(crate) mod session;
+mod tls;
 
 async fn connect_tcp_inner(addr: SocketAddr, timeout_val: Duration) -> Result<TcpStream> {
     let tcp_stream = timeout(timeout_val, TcpStream::connect(addr))
